Moved TransparentFunctors to Utility/String
