+++ src/Bin/CodeGen/CodeGenMap.h
@@ -11,7 +11,7 @@
 
 #include "Utility/Workaround/ToUnderlying.h"
 #include "Utility/Format.h"
-#include "Utility/TransparentFunctors.h"
+#include "Utility/String/TransparentFunctors.h"
 
 #include "CodeGenFunctions.h"
