+++ src/Utility/String/TransparentFunctors.h
